+++ packages/breadboard/src/data/index.ts
@@ -4,48 +4,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { asBase64, isStoredData } from "./data.js";
-import {
-  DataStore,
-  InlineDataCapabilityPart,
-  StoredDataCapabilityPart,
-} from "./types.js";
+import { SimpleDataStore } from "./simple.js";
+import { DataStore } from "./types.js";
 
 export const createDataStore = (): DataStore => {
   return new SimpleDataStore();
 };
 
-class SimpleDataStore implements DataStore {
-  async store(data: Blob): Promise<StoredDataCapabilityPart> {
-    const handle = URL.createObjectURL(data);
-    const mimeType = data.type;
-    return {
-      storedData: { handle, mimeType },
-    };
-  }
-  async retrieve(
-    storedData: StoredDataCapabilityPart
-  ): Promise<InlineDataCapabilityPart> {
-    const raw = await this.retrieveAsBlob(storedData);
-    const mimeType = storedData.storedData.mimeType;
-    const data = await asBase64(raw);
-    return { inlineData: { mimeType, data } };
-  }
+export { inflateData, deflateData } from "./inflate-deflate.js";
 
-  async retrieveAsBlob(storedData: StoredDataCapabilityPart): Promise<Blob> {
-    if (!isStoredData(storedData)) {
-      throw new Error("Invalid stored data");
-    }
-    const { handle } = storedData.storedData;
-    const response = await fetch(handle);
-    return await response.blob();
-  }
-
-  async retrieveAsURL(storedData: StoredDataCapabilityPart): Promise<string> {
-    if (!isStoredData(storedData)) {
-      throw new Error("Invalid stored data");
-    }
-    const { handle } = storedData.storedData;
-    return handle;
-  }
-}
+export { isDataCapability, asBlob } from "./common.js";
